requirements.txt
torch
transformers
lightning
hydra-core
omegaconf
datasets
tqdm
numpy
scipy
numba
faiss-cpu
